Major updates to pymc branch, which I will now switch to become the default.
Programmatically, this includes a major change away from modelling the period aliases or gaps of the mono/duo/trio/etc within a `pymc` model (i.e. linking the model fit parameters to the derived period and marginalised period parameters for each sample), to now computing a period-ambiguous `pymc` model (which is somewhat quicked) and then calculating the derived probability distribution a posteriori. This does mean that, for very discontinuous models, there may be some disconnect between the transit model and the derived period distribution (i.e. in a case where there is one very likely & very short period alias you may want a connection between derived period posterior to push the model deeper towards that in the transit fit). In typical TESS cases - monotransits with relatively continuous period spaces, or duotransits with many aliases stacked together, then the impact is minimal.
One way to implement this was custom `exoplanet` orbit cases which allow for Single and Duo transit cases (effectively just in the phase calculation).

Other updates include:
- Handful of new examples including ambiguous (i.e. trio+) cases
- Including TTVs in ambiguous cases with >2 transits
- Removing support for RVs (they never really worked anyway)
- Removing support for in-transit spline detrending.

And of course, being in `pymc`, we no longer have as many wholly disgusting dependencies due to numpy and theano and that whole mess.